FixtureDataItem -> LookupTableRow in linked_domain app

+++ corehq/apps/linked_domain/updates.py
@@ -5,6 +5,8 @@ from functools import partial
 from django.utils.translation import gettext as _
 from django.db import transaction
 
+from dimagi.utils.chunked import chunked
+
 from corehq.apps.data_interfaces.models import (
     AutomaticUpdateRule, CaseRuleAction, CaseRuleCriteria,
     ClosedParentDefinition, CustomActionDefinition,
@@ -27,7 +29,7 @@ from corehq.apps.integration.models import (
     HmacCalloutSettings,
 )
 from corehq.apps.fixtures.dbaccessors import delete_fixture_items_for_data_type
-from corehq.apps.fixtures.models import FixtureDataItem, LookupTable
+from corehq.apps.fixtures.models import LookupTable, LookupTableRow
 from corehq.apps.fixtures.upload.run_upload import clear_fixture_quickcache
 from corehq.apps.fixtures.utils import clear_fixture_cache
 from corehq.apps.linked_domain.const import (
@@ -219,8 +221,10 @@ def update_fixture(domain_link, tag):
     try:
         linked_data_type = LookupTable.objects.by_domain_tag(
             domain_link.linked_domain, master_data_type.tag)
+        is_existing_table = True
     except LookupTable.DoesNotExist:
         linked_data_type = LookupTable(domain=domain_link.linked_domain)
+        is_existing_table = False
     for field in LookupTable._meta.fields:
         if field.attname not in ["id", "domain"]:
             value = getattr(master_data_type, field.attname)
@@ -229,14 +233,23 @@ def update_fixture(domain_link, tag):
     clear_fixture_quickcache(domain_link.linked_domain, [linked_data_type])
 
     # Re-create relevant data items
-    delete_fixture_items_for_data_type(domain_link.linked_domain, linked_data_type._migration_couch_id)
-    for master_item in master_results["data_items"]:
-        doc = master_item.to_json()
-        del doc["_id"]
-        del doc["_rev"]
-        doc["domain"] = domain_link.linked_domain
-        doc["data_type_id"] = linked_data_type._migration_couch_id
-        FixtureDataItem.wrap(doc).save()
+    if is_existing_table:
+        LookupTableRow.objects.filter(
+            domain=domain_link.linked_domain,
+            table_id=linked_data_type.id
+        ).delete()
+        delete_fixture_items_for_data_type(domain_link.linked_domain, linked_data_type._migration_couch_id)
+    ignore_fields = {"id", "domain", "table", "table_id"}
+    row_fields = [field.attname
+        for field in LookupTableRow._meta.fields
+        if field.attname not in ignore_fields]
+    rows = (LookupTableRow(
+        domain=domain_link.linked_domain,
+        table_id=linked_data_type.id,
+        **{f: getattr(master_item, f) for f in row_fields}
+    ) for master_item in master_results["data_items"])
+    for chunk in chunked(rows, 1000, list):
+        LookupTableRow.objects.bulk_create(chunk)
 
     clear_fixture_cache(domain_link.linked_domain)
